# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Telegram Pulse API
  version: 1.0.0
  summary: Curated REST contract for UU integration on top of custom auth.
  description: |
    Multi-tenant API for Telegram analytics, ads intelligence, trackers, rankings,
    mini apps, account management, team management, billing, and exports.
servers:
  - url: https://api.telegrampulse.example.com
    description: Production
  - url: https://staging-api.telegrampulse.example.com
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /v1.0/channels:
    get:
      tags: [Channels]
      operationId: listChannels
      summary: Search and filter channels catalog.
      parameters:
        - $ref: '#/components/parameters/QueryParam'
        - name: country_code
          in: query
          schema: { type: string, minLength: 2, maxLength: 2 }
        - name: category_slug
          in: query
          schema: { type: string }
        - name: size_bucket
          in: query
          schema:
            type: string
            enum: [small, medium, large, huge]
        - name: er_min
          in: query
          schema: { type: number, minimum: 0, maximum: 100 }
        - name: er_max
          in: query
          schema: { type: number, minimum: 0, maximum: 100 }
        - name: status
          in: query
          schema:
            type: string
            enum: [normal, verified, scam, restricted, deleted]
        - name: verified
          in: query
          schema: { type: boolean }
        - name: scam
          in: query
          schema: { type: boolean }
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [subscribers, growth_24h, growth_7d, growth_30d, engagement_rate, updated_at]
            default: subscribers
        - $ref: '#/components/parameters/SortOrderParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelListEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1.0/accounts/{accountId}/channels:
    get:
      tags: [Account Channels]
      operationId: listAccountChannels
      summary: List channels attached to account.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/CursorParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountChannelListEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [Account Channels]
      operationId: addAccountChannel
      summary: Add channel to account tracking list.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddAccountChannelRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountChannelEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/channels/insights:
    get:
      tags: [Account Channels]
      operationId: getAccountChannelInsights
      summary: Aggregate account channel insights.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountChannelInsightsEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/channels/{channelId}/verification:
    post:
      tags: [Account Channels]
      operationId: createChannelVerificationRequest
      summary: Start channel verification flow.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - $ref: '#/components/parameters/ChannelIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationRequestCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRequestEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/channels/{channelId}/verification/{requestId}/confirm:
    post:
      tags: [Account Channels]
      operationId: confirmChannelVerificationRequest
      summary: Confirm verification request.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - $ref: '#/components/parameters/ChannelIdPath'
        - $ref: '#/components/parameters/VerificationRequestIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerificationConfirmRequest' }
      responses:
        '200':
          description: Confirmed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationRequestEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1.0/accounts/{accountId}/api-keys:
    get:
      tags: [API Keys]
      operationId: listApiKeys
      summary: List account API keys (masked).
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyListEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      tags: [API Keys]
      operationId: createApiKey
      summary: Create API key (secret returned once).
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiKeyCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyCreateEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1.0/accounts/{accountId}/api-keys/{apiKeyId}/rotate:
    post:
      tags: [API Keys]
      operationId: rotateApiKey
      summary: Rotate API key secret.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - $ref: '#/components/parameters/ApiKeyIdPath'
      responses:
        '200':
          description: Rotated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyCreateEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1.0/accounts/{accountId}/api-keys/{apiKeyId}:
    delete:
      tags: [API Keys]
      operationId: deleteApiKey
      summary: Revoke API key.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - $ref: '#/components/parameters/ApiKeyIdPath'
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1.0/accounts/{accountId}/api-usage:
    get:
      tags: [API Keys]
      operationId: getApiUsage
      summary: Get API usage aggregates for account.
      parameters:
        - $ref: '#/components/parameters/AccountPath'
        - $ref: '#/components/parameters/AccountHeader'
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiUsageEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AccountHeader:
      name: X-Account-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Account context. Must match path accountId when both are present.
    AccountPath:
      name: accountId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ChannelIdPath:
      name: channelId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    VerificationRequestIdPath:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ApiKeyIdPath:
      name: apiKeyId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    QueryParam:
      name: q
      in: query
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
    CursorParam:
      name: cursor
      in: query
      schema: { type: string }
    SortOrderParam:
      name: sort_order
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: unauthorized
              message: Missing or invalid bearer token.
              details: []
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: forbidden
              message: You do not have access to this account resource.
              details: []
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: not_found
              message: Resource was not found.
              details: []
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            error:
              code: validation_error
              message: Request validation failed.
              details:
                - field: er_min
                  issue: er_min must be less than or equal to er_max
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
    Page:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    Meta:
      type: object
      additionalProperties: true
...
    ChannelListItem:
      type: object
      properties:
        channel_id: { type: string, format: uuid }
        name: { type: string }
        username: { type: [string, 'null'] }
        subscribers: { type: integer }
        growth_24h: { type: [number, 'null'] }
        growth_7d: { type: [number, 'null'] }
        growth_30d: { type: [number, 'null'] }
        engagement_rate: { type: [number, 'null'] }
        category_slug: { type: [string, 'null'] }
        category_name: { type: [string, 'null'] }
        country_code: { type: [string, 'null'] }
        status: { type: string }
        verified: { type: boolean }
        scam: { type: boolean }
...
    AccountChannel:
      type: object
      properties:
        account_id: { type: string, format: uuid }
        channel_id: { type: string, format: uuid }
        alias_name: { type: [string, 'null'] }
        monitoring_enabled: { type: boolean }
        is_favorite: { type: boolean }
        added_at: { type: string, format: date-time }
    AddAccountChannelRequest:
      type: object
      required: [channel_id]
      properties:
        channel_id: { type: string, format: uuid }
        alias_name: { type: string }
        monitoring_enabled: { type: boolean, default: true }
        is_favorite: { type: boolean, default: false }
    AccountChannelInsights:
      type: object
      properties:
        total_subscribers: { type: integer }
        total_views: { type: integer }
        avg_engagement_rate: { type: number }
        channels_count: { type: integer }
    VerificationRequest:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        channel_id: { type: string, format: uuid }
        verification_code: { type: string }
        verification_method: { type: string }
        status: { type: string }
        requested_at: { type: string, format: date-time }
        confirmed_at: { type: [string, 'null'], format: date-time }
        expires_at: { type: string, format: date-time }
    VerificationRequestCreateRequest:
      type: object
      properties:
        verification_method:
          type: string
          enum: [description_code, bot_admin, manual_review]
          default: description_code
    VerificationConfirmRequest:
      type: object
      properties:
        evidence:
          type: object
          additionalProperties: true
...
    ApiKeyListItem:
      type: object
      properties:
        api_key_id: { type: string, format: uuid }
        name: { type: string }
        key_prefix: { type: string }
        scopes:
          type: array
          items: { type: string }
        rate_limit_per_hour: { type: integer }
        created_at: { type: string, format: date-time }
        last_used_at: { type: [string, 'null'], format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
    ApiKeyCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        scopes:
          type: array
          items: { type: string }
        rate_limit_per_hour: { type: integer, minimum: 1 }
    ApiKeyCreateResponse:
      type: object
      properties:
        api_key:
          $ref: '#/components/schemas/ApiKeyListItem'
        secret:
          type: string
          description: Returned once only at create/rotate time.
    ApiUsage:
      type: object
      properties:
        total_requests: { type: integer }
        error_rate: { type: number }
        avg_latency_ms: { type: number }
        by_day:
          type: array
          items:
            type: object
            properties:
              date: { type: string, format: date }
              requests: { type: integer }
              errors: { type: integer }
...
    ChannelListEnvelope:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ChannelListItem' }
        page: { $ref: '#/components/schemas/Page' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    AccountChannelEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/AccountChannel' }
        meta: { $ref: '#/components/schemas/Meta' }
    AccountChannelListEnvelope:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AccountChannel' }
        page: { $ref: '#/components/schemas/Page' }
        meta: { $ref: '#/components/schemas/Meta' }
    AccountChannelInsightsEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/AccountChannelInsights' }
        meta: { $ref: '#/components/schemas/Meta' }
    VerificationRequestEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/VerificationRequest' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    ApiKeyListEnvelope:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ApiKeyListItem' }
        meta: { $ref: '#/components/schemas/Meta' }
    ApiKeyCreateEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/ApiKeyCreateResponse' }
        meta: { $ref: '#/components/schemas/Meta' }
    ApiUsageEnvelope:
      type: object
      properties:
        data: { $ref: '#/components/schemas/ApiUsage' }
        meta: { $ref: '#/components/schemas/Meta' }
...